refactor(talk): isolate prepared session target type
Keep the prepared identity contract independent of storage runtime imports. Migrate all internal type consumers and remove the old runtime-module type export, breaking the static import cycle without changing runtime values.

Full architecture checks, core and test types, 209 Talk owner tests, build, erased-JavaScript comparison, and fresh scoped review pass.
