+++ src/gateway/server-methods/shared-types.ts
@@ -398,7 +398,7 @@ export type GatewayRequestOptions = {
 
 /** Commit-time guard captured by the pre-dispatch session participation check. */
 export type SessionMutationAuthorization = {
-  talkSessionTarget?: import("../talk-session-target.js").PreparedTalkSessionTarget;
+  talkSessionTarget?: import("../talk-session-target.types.js").PreparedTalkSessionTarget;
   assertCurrent: () => void;
   assertTargetCurrent: (target: {
     sessionKey: string;
